BIG update
got the crafting minigame going and working with the regular minigames. ive still yet to implement it fully but i might not have to. honestly i just have to look at the code and determine it. all thats left now is to get the animations going for the minigame and then i can start working on the prescriptionNPCs